Completed transition between sub-pages of About Us and What We Do V1
The transition part might be a little confusing, especially on the direction. Here's a little clarification on that matter:
When the user wants to navigate to the right (where their hand goes from right to left) or right arrow key is pressed, on the animation object's perspective, it is going from right to left, hence the "SwipeLeft" animation clip is called.

When the user wants to navigate to the left (where their hand goes from left to right) or left arrow key is pressed, on the animation object's perspective, it is going from left to right, hence the "SwipeRight" animation clip is called.